requirements.txt
selenium==4.39.0
pytest==9.0.2
allure-pytest==2.15.3
webdriver-manager==4.0.2